Modified structure of components
