+++ packages/core/src/components/PinInput/index.tsx
@@ -1,3 +1,3 @@
 export { default as PinInput } from "./PinInput";
-export { default as PinInputCell } from "./PinInputCell";
+export { default as CustomPinInputCell } from "./CustomPinInputCell";
 export { default as PinInputText } from "./PinInputText";
