+++ packages/ui/src/index.tsx
@@ -60,8 +60,8 @@ export {
   VideoPlayer,
   VideoPlayerRef,
   PinInput,
-  PinInputCell,
-  PinInputText,
+  CustomPinInputCell,
+  CustomPinInputText,
   AspectRatio,
   HStack,
   VStack,
